Bug fixes and big data file removal

+++ Minotaur-SGX/adaptWordCount/App/utils.hpp
@@ -14,7 +14,7 @@
 struct Arguments{
     const char * ip; 
     int id;
-    int windowSize;
+    float windowSize;
     int prev_stage;
     bool multiout = false; 
     std::string name;
